backend/services/catalog/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose[cryptography]
sqlalchemy>=2.0
asyncpg
alembic
